pyproject.toml
[project]
name = "fpga-sr-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
